+++ playtime-0.6.0-ts/src/api-routes.js
@@ -1,8 +1,8 @@
 import { usersApi } from "./api/users-api.js";
 
 export const apiRoutes = [
-  { method: "GET", path: "/api/users", config: usersApi.find },
-  { method: "POST", path: "/api/users", config: usersApi.create },
-  { method: "DELETE", path: "/api/users", config: usersApi.deleteAll },
-  { method: "GET", path: "/api/users/{id}", config: usersApi.findOne },
+  { method: "GET", path: "/api/userApi", config: usersApi.find },
+  { method: "POST", path: "/api/userApi", config: usersApi.create },
+  { method: "DELETE", path: "/api/userApi", config: usersApi.deleteAll },
+  { method: "GET", path: "/api/userApi/{id}", config: usersApi.findOne },
 ];