+++ playtime-0.6.0-ts/test/models/users-model-test.js
@@ -19,7 +19,7 @@ suite("User Model tests", () => {
     assertSubset(maggie, newUser);
   });
 
-  test("delete all users", async () => {
+  test("delete all userApi", async () => {
     let returnedUsers = await db.userStore.getAllUsers();
     assert.equal(returnedUsers.length, 3);
     await db.userStore.deleteAll();