+++ playtime-0.6.0/src/models/json/track-json-store.js
@@ -2,7 +2,7 @@ import { v4 } from "uuid";
 // eslint-disable-next-line import/no-unresolved
 import { JSONFile, Low } from "lowdb";
 
-const db = new Low(new JSONFile("./src/models/json/tracks.json"));
+const db = new Low(new JSONFile("./src/models/json/trackApi.json"));
 db.data = { tracks: [] };
 
 export const trackJsonStore = {