+++ playtime-0.7.0/src/api/playlist-api.js
@@ -2,7 +2,7 @@ import Boom from "@hapi/boom";
 import { PlaylistSpec } from "../models/joi-schemas.js";
 import { db } from "../models/db.js";
 
-export const PlaylistApi = {
+export const playlistApi = {
   find: {
     auth: false,
     handler: async function (request, h) {