+++ playtime-0.7.0/test/api/tracks-api-test.js
@@ -15,7 +15,7 @@ suite("Track API tests", () => {
   test("create track", async () => {
   });
 
-  test("create Multiple tracks", async () => {
+  test("create Multiple trackApi", async () => {
   });
 
   test("Delete Track", async () => {