+++ playtime-0.8.0/src/api-routes.js
@@ -1,22 +1,22 @@
-import { Users } from "./api/users.js";
-import { Playlists } from "./api/playlists.js";
-import { Tracks } from "./api/tracks.js";
+import { userApi } from "./api/user-api.js";
+import { playlistApi } from "./api/playlist-api.js";
+import { trackApi } from "./api/track-api.js";
 
 export const apiRoutes = [
-  { method: "GET", path: "/api/users", config: Users.find },
-  { method: "POST", path: "/api/users", config: Users.create },
-  { method: "DELETE", path: "/api/users", config: Users.deleteAll },
-  { method: "GET", path: "/api/users/{id}", config: Users.findOne },
+  { method: "GET", path: "/api/users", config: userApi.find },
+  { method: "POST", path: "/api/users", config: userApi.create },
+  { method: "DELETE", path: "/api/users", config: userApi.deleteAll },
+  { method: "GET", path: "/api/users/{id}", config: userApi.findOne },
 
-  { method: "POST", path: "/api/playlists", config: Playlists.create },
-  { method: "DELETE", path: "/api/playlists", config: Playlists.deleteAll },
-  { method: "GET", path: "/api/playlists", config: Playlists.find },
-  { method: "GET", path: "/api/playlists/{id}", config: Playlists.findOne },
-  { method: "DELETE", path: "/api/playlists/{id}", config: Playlists.deleteOne },
+  { method: "POST", path: "/api/playlists", config: playlistApi.create },
+  { method: "DELETE", path: "/api/playlists", config: playlistApi.deleteAll },
+  { method: "GET", path: "/api/playlists", config: playlistApi.find },
+  { method: "GET", path: "/api/playlists/{id}", config: playlistApi.findOne },
+  { method: "DELETE", path: "/api/playlists/{id}", config: playlistApi.deleteOne },
 
-  { method: "GET", path: "/api/tracks", config: Tracks.find },
-  { method: "GET", path: "/api/tracks/{id}", config: Tracks.findOne },
-  { method: "POST", path: "/api/playlists/{id}/tracks", config: Tracks.create },
-  { method: "DELETE", path: "/api/tracks", config: Tracks.deleteAll },
-  { method: "DELETE", path: "/api/tracks/{id}", config: Tracks.deleteOne },
+  { method: "GET", path: "/api/tracks", config: trackApi.find },
+  { method: "GET", path: "/api/tracks/{id}", config: trackApi.findOne },
+  { method: "POST", path: "/api/playlists/{id}/tracks", config: trackApi.create },
+  { method: "DELETE", path: "/api/tracks", config: trackApi.deleteAll },
+  { method: "DELETE", path: "/api/tracks/{id}", config: trackApi.deleteOne },
 ];