+++ playtime-0.8.0/src/api/user-api.js
@@ -1,10 +1,12 @@
 import Boom from "@hapi/boom";
 import { db } from "../models/db.js";
+import { validationError } from "../utils/logger.js";
+import { IdSpec, UserArray, UserSpec } from "../models/joi-schemas.js";
 
-export const usersApi = {
+export const userApi = {
   find: {
     auth: false,
-    handler: async function (request, h) {
+    handler: async function(request, h) {
       try {
         const users = await db.userStore.getAllUsers();
         return users;
@@ -12,11 +14,15 @@ export const usersApi = {
         return Boom.serverUnavailable("Database Error");
       }
     },
+    tags: ["api"],
+    description: "Get all userApi",
+    notes: "Returns details of all userApi",
+    response: { schema: UserArray, failAction: validationError },
   },
 
   findOne: {
     auth: false,
-    handler: async function (request, h) {
+    handler: async function(request, h) {
       try {
         const user = await db.userStore.getUserById(request.params.id);
         if (!user) {
@@ -27,11 +33,16 @@ export const usersApi = {
         return Boom.serverUnavailable("No User with this id");
       }
     },
+    tags: ["api"],
+    description: "Get a specific user",
+    notes: "Returns user details",
+    validate: { params: { id: IdSpec }, failAction: validationError },
+    response: { schema: UserSpec, failAction: validationError },
   },
 
   create: {
     auth: false,
-    handler: async function (request, h) {
+    handler: async function(request, h) {
       try {
         const user = await db.userStore.addUser(request.payload);
         if (user) {
@@ -42,11 +53,16 @@ export const usersApi = {
         return Boom.serverUnavailable("Database Error");
       }
     },
+    tags: ["api"],
+    description: "Create a User",
+    notes: "Returns the newly created user",
+    validate: { payload: UserSpec, failAction: validationError },
+    response: { schema: UserSpec, failAction: validationError },
   },
 
   deleteAll: {
     auth: false,
-    handler: async function (request, h) {
+    handler: async function(request, h) {
       try {
         await db.userStore.deleteAll();
         return h.response().code(204);
@@ -54,5 +70,8 @@ export const usersApi = {
         return Boom.serverUnavailable("Database Error");
       }
     },
+    tags: ["api"],
+    description: "Delete all userApi",
+    notes: "All userApi removed from Playtime",
   },
 };