+++ playtime-0.8.0/test/api/tracks-api-test.js
@@ -1,10 +1,9 @@
 import { assert } from "chai";
-import { isSubset } from "./test-utils.js";
-import * as fixtures from "./fixtures.json";
+import { assertSubset } from "../test-utils.js";
 import { playtimeService } from "./playtime-service.js";
+import { maggie, mozart, testPlaylists, testTracks, concerto } from "../fixtures.js";
 
 suite("Track API tests", () => {
-  const { maggie, beethoven, sonataNo3, testTracks } = fixtures.default;
   let user = null;
   let beethovenSonatas = null;
 
@@ -13,15 +12,15 @@ suite("Track API tests", () => {
     await playtimeService.deleteAllUsers();
     await playtimeService.deleteAllTracks();
     user = await playtimeService.createUser(maggie);
-    beethoven.userid = user._id;
-    beethovenSonatas = await playtimeService.createPlaylist(beethoven);
+    mozart.userid = user._id;
+    beethovenSonatas = await playtimeService.createPlaylist(mozart);
   });
 
   teardown(async () => {});
 
   test("create track", async () => {
-    const returnedTrack = await playtimeService.createTrack(beethovenSonatas._id, sonataNo3);
-    assert(isSubset(sonataNo3, returnedTrack), "returned track must be a superset of new track");
+    const returnedTrack = await playtimeService.createTrack(beethovenSonatas._id, concerto);
+    assertSubset(concerto, returnedTrack);
   });
 
   test("create Multiple tracks", async () => {
@@ -34,7 +33,7 @@ suite("Track API tests", () => {
     for (let i = 0; i < returnedTracks.length; i += 1) {
       // eslint-disable-next-line no-await-in-loop
       const track = await playtimeService.getTrack(returnedTracks[i]._id);
-      assert(isSubset(track, returnedTracks[i]), "returned track must be a superset of new track");
+      assertSubset(track, returnedTracks[i]);
     }
   });
 
@@ -61,7 +60,7 @@ suite("Track API tests", () => {
     const returnedPlaylist = await playtimeService.getPlaylist(beethovenSonatas._id);
     assert.equal(returnedPlaylist.tracks.length, testTracks.length);
     for (let i = 0; i < testTracks.length; i += 1) {
-      assert(isSubset(testTracks[i], returnedPlaylist.tracks[i]));
+      assertSubset(testTracks[i], returnedPlaylist.tracks[i]);
     }
   });
 });