+++ playtime-0.4.0/CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ## version 0.3.0
 
-- support delete playlist + tracks
+- support delete playlist + trackApi
 - introduce .env for secrets
 - include Joi schema validation for signup form
 - incorporate JSON data stores
@@ -19,9 +19,9 @@
 - about controller + view added
 - introduce track model
 - playlist controller + view & partials added
-- support adding tracks to playlists
+- support adding trackApi to playlists
 - introduce sessions
-- associate playlists with users
+- associate playlists with userApi
 
 ## version 0.1.0